Refactored out `cursors_utils.py`
